- renamed db_backup and db_cronjob to adhere to naming conventions
